# src/docs/swagger.yaml
basePath: /
definitions:
...
  types.BadRequestErrorResponseStruct:
    properties:
      message:
        example: Bad Request
        type: string
      statusCode:
        example: 400
        type: integer
      success:
        example: false
        type: boolean
    type: object
  types.CategoryQuizResponseDTOStruct:
    properties:
      id:
        example: d27b21ab-6177-4159-9e13-15dc50ffed29
        type: string
      name:
        example: General Knowledge
        type: string
    type: object
...
  types.CreateQuizQuestionChoiceStruct:
    properties:
      content:
        example: Paris
        type: string
      is_correct:
        example: true
        type: boolean
    type: object
  types.CreateQuizQuestionsStruct:
    properties:
      choices:
        items:
          $ref: '#/definitions/types.CreateQuizQuestionChoiceStruct'
        type: array
      content:
        example: What is the capital of France?
        type: string
    required:
    - choices
    - content
    type: object
  types.CreateQuizRequestBody:
    properties:
      category_id:
        example: d27b21ab-6177-4159-9e13-15dc50ffed29
        type: string
      name:
        example: Sample Quiz
        type: string
      questions:
        items:
          $ref: '#/definitions/types.CreateQuizQuestionsStruct'
        type: array
    required:
    - category_id
    - name
    - questions
    type: object
  types.CreateQuizSuccessResponseStruct:
    properties:
      data:
        $ref: '#/definitions/types.QuizResponseDTO'
      statusCode:
        example: 201
        type: integer
      success:
        example: true
        type: boolean
    type: object
  types.ForbiddenErrorResponseStruct:
    properties:
      message:
        example: Forbidden
        type: string
      statusCode:
        example: 403
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  types.GetQuizSuccessResponseStruct:
    properties:
      data:
        $ref: '#/definitions/types.QuizDetailedResponseDTO'
      statusCode:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
  types.GetQuizzesSuccessResponseStruct:
    properties:
      data:
        items:
          $ref: '#/definitions/types.QuizResponseDTO'
        type: array
      statusCode:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
...
  types.InternalServerErrorResponseStruct:
    properties:
      message:
        default: Internal Server Error
        type: string
      statusCode:
        default: 500
        type: integer
      success:
        default: false
        type: boolean
    type: object
...
  types.NotFoundErrorResponseStruct:
    properties:
      message:
        example: Not Found
        type: string
      statusCode:
        example: 404
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  types.QuizDetailedResponseDTO:
    properties:
      category:
        $ref: '#/definitions/types.CategoryQuizResponseDTOStruct'
      category_id:
        example: d27b21ab-6177-4159-9e13-15dc50ffed29
        type: string
      created_at:
        example: "2025-10-22T19:01:58.778079424Z"
        type: string
      created_by:
        example: 0fde5216-1bab-41f6-bd90-4c3f088ee91f
        type: string
      curator_pick:
        example: false
        type: boolean
      games_played:
        example: 0
        type: integer
      id:
        example: 4fdb53f5-74d2-4d0e-8267-43f893a51aca
        type: string
      likes:
        example: 0
        type: integer
      name:
        example: Sample Quiz
        type: string
      questions:
        items:
          $ref: '#/definitions/types.QuizQuestionResponseDTO'
        type: array
      updated_at:
        example: "2025-10-22T19:01:58.778079424Z"
        type: string
      user:
        $ref: '#/definitions/types.UserQuizResponseDTOStruct'
    type: object
...
  types.QuizResponseDTO:
    properties:
      category:
        $ref: '#/definitions/types.CategoryQuizResponseDTOStruct'
      category_id:
        example: d27b21ab-6177-4159-9e13-15dc50ffed29
        type: string
      created_at:
        example: "2025-10-22T19:01:58.778079424Z"
        type: string
      created_by:
        example: 0fde5216-1bab-41f6-bd90-4c3f088ee91f
        type: string
      curator_pick:
        example: false
        type: boolean
      games_played:
        example: 0
        type: integer
      id:
        example: 4fdb53f5-74d2-4d0e-8267-43f893a51aca
        type: string
      likes:
        example: 0
        type: integer
      name:
        example: Sample Quiz
        type: string
      updated_at:
        example: "2025-10-22T19:01:58.778079424Z"
        type: string
      user:
        $ref: '#/definitions/types.UserQuizResponseDTOStruct'
    type: object
...
  types.SuccessResponseStruct:
    properties:
      message:
        example: OK
        type: string
      statusCode:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
...
  types.UpdateQuizRequestBody:
    properties:
      category_id:
        example: d27b21ab-6177-4159-9e13-15dc50ffed29
        type: string
      name:
        example: Sample Quiz
        type: string
    type: object
...
  types.UserQuizResponseDTOStruct:
    properties:
      id:
        example: 0fde5216-1bab-41f6-bd90-4c3f088ee91f
        type: string
      name:
        example: John Doe
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service for IntelliQuiz's web application purposes.
  license:
    name: MIT License
    url: https://mit-license.org/
  title: IntelliQuiz API
  version: "1.0"
paths:
...
  /me/quizzes:
    get:
      description: Retrieve a list of quizzes created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetQuizzesSuccessResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Get own quizzes
      tags:
      - quizzes
...
  /quizzes:
    get:
      description: Retrieve a list of all quizzes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetQuizzesSuccessResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Get all quizzes
      tags:
      - quizzes
    post:
      description: Create a new quiz
      parameters:
      - description: Create Quiz Request Body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.CreateQuizRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.CreateQuizSuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Create a new quiz
      tags:
      - quizzes
  /quizzes/{quizId}:
    delete:
      description: Delete a quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.NotFoundErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Delete a quiz by ID
      tags:
      - quizzes
    get:
      description: Retrieve a quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetQuizSuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.NotFoundErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Get a quiz by ID
      tags:
      - quizzes
    patch:
      consumes:
      - application/json
      description: Update a quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Quiz Request Body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateQuizRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.NotFoundErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Update a quiz by ID
      tags:
      - quizzes
  /quizzes/{quizId}/dislike:
    post:
      description: Dislike a quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.NotFoundErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Dislike a quiz by ID
      tags:
      - quizzes
  /quizzes/{quizId}/like:
    post:
      description: Like a quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.BadRequestErrorResponseStruct'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ForbiddenErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.NotFoundErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.InternalServerErrorResponseStruct'
      summary: Like a quiz by ID
      tags:
      - quizzes
...
swagger: "2.0"